refactor(container): extract Android/Playwright MCP helpers to satisfy max-lines lint
Move the Android and Playwright MCP config/sidecar builders into sibling
modules so grok.ts, gemini.ts, and docker-compose.ts stay under the 300-line
max-lines budget enforced by CI. Pure refactor — rendered output is unchanged
(72 container tests still pass).

REF: issue-436
