+++ packages/container/src/core/templates/docker-compose.ts
@@ -7,6 +7,8 @@ import {
   type TemplateConfig
 } from "../domain.js"
 import type { ResolvedComposeResourceLimits } from "../resource-limits.js"
+import { buildAndroidFragments } from "./docker-compose-android.js"
+import { buildPlaywrightFragments } from "./docker-compose-playwright.js"
 
 type ComposeFragments = {
   readonly networkMode: TemplateConfig["dockerNetworkMode"]
@@ -29,21 +31,6 @@ type ComposeFragments = {
   readonly maybeBootstrapMounts: string
   readonly forkRepoUrl: string
 }
-
-type PlaywrightFragments = Pick<
-  ComposeFragments,
-  | "maybeDependsOn"
-  | "maybeDockerSocketMount"
-  | "maybePlaywrightEnv"
-  | "maybeBrowserVolume"
->
-
-type AndroidFragments = Pick<
-  ComposeFragments,
-  | "maybeAndroidEnv"
-  | "maybeAndroidService"
-  | "maybeAndroidVolume"
->
 
 type AuthEnvFragments = Pick<
   ComposeFragments,
@@ -122,13 +109,6 @@ const renderBootstrapMounts = (): string => `      - ${bootstrapVolumeKey}:/opt/
 
 const renderYamlSingleQuoted = (value: string): string => `'${value.replaceAll("'", "''")}'`
 
-const renderOptionalDockerSocketMount = (
-  shouldEnableLocalDockerSocket: boolean
-): string =>
-  shouldEnableLocalDockerSocket
-    ? `      - /var/run/docker.sock:/var/run/docker.sock`
-    : ""
-
 const renderEnvFiles = (config: TemplateConfig): string =>
   `    env_file:\n      - ${renderYamlSingleQuoted(config.envGlobalPath)}\n      - ${
     renderYamlSingleQuoted(
@@ -160,102 +140,6 @@ const buildAgentEnvFragments = (config: TemplateConfig): AgentEnvFragments => ({
   maybeAgentModeEnv: renderAgentModeEnv(config.agentMode),
   maybeAgentAutoEnv: renderAgentAutoEnv(config.agentAuto)
 })
-
-const renderBrowserLimitEnv = (
-  key: string,
-  value: number | string | undefined
-): string => `      ${key}: "\${${key}:-${value ?? ""}}"\n`
-
-const buildPlaywrightFragments = (
-  config: TemplateConfig,
-  resourceLimits: ResolvedComposeResourceLimits | undefined,
-  options: DockerComposeRenderOptions
-): PlaywrightFragments => {
-  if (!config.enableMcpPlaywright) {
-    return {
-      maybeDependsOn: "",
-      maybeDockerSocketMount: "",
-      maybePlaywrightEnv: "",
-      maybeBrowserVolume: ""
-    }
-  }
-
-  const browserContainerName = `${config.containerName}-browser`
-  const browserVolumeName = `${config.volumeName}-browser`
-  const browserImageName = `${browserContainerName}:docker-git-browser`
-
-  return {
-    maybeDependsOn: "",
-    maybeDockerSocketMount: renderOptionalDockerSocketMount(
-      options.enableLocalDockerSocket
-    ),
-    maybePlaywrightEnv:
-      `      MCP_PLAYWRIGHT_ENABLE: "1"\n      DOCKER_GIT_PROJECT_CONTAINER_NAME: "${config.containerName}"\n      DOCKER_GIT_BROWSER_CONTAINER_NAME: "${browserContainerName}"\n      DOCKER_GIT_BROWSER_IMAGE_NAME: "${browserImageName}"\n      DOCKER_GIT_BROWSER_VOLUME_NAME: "${browserVolumeName}"\n${
-        renderBrowserLimitEnv(
-          "DOCKER_GIT_BROWSER_CPU_LIMIT",
-          resourceLimits?.cpuLimit
-        )
-      }${renderBrowserLimitEnv("DOCKER_GIT_BROWSER_RAM_LIMIT", resourceLimits?.ramLimit)}`,
-    maybeBrowserVolume: `  ${browserVolumeName}:`
-  }
-}
-
-// CHANGE: render an Android emulator sidecar service mirroring the Playwright MCP wiring
-// WHY: issue-436 asks to connect mcp-android "the same way" Playwright MCP works, exposing
-//      a docker-android emulator as a service reachable over ADB for the mobile-mcp server
-// QUOTE(ТЗ): "Подключить mcp-android так же как работает MCP PLAYRIGHT"
-// REF: issue-436
-// SOURCE: https://github.com/budtmo/docker-android
-// PURITY: CORE
-// INVARIANT: only emitted when config.enableMcpAndroid === true; image/ports are env-overridable
-const defaultAndroidEmulatorImage = "budtmo/docker-android:emulator_14.0"
-
-const buildAndroidFragments = (
-  config: TemplateConfig,
-  resourceLimits: ResolvedComposeResourceLimits | undefined
-): AndroidFragments => {
-  if (config.enableMcpAndroid !== true) {
-    return {
-      maybeAndroidEnv: "",
-      maybeAndroidService: "",
-      maybeAndroidVolume: ""
-    }
-  }
-
-  const androidContainerName = `${config.containerName}-android`
-  const androidVolumeName = `${config.volumeName}-android`
-  const androidImageRef = `\${DOCKER_GIT_ANDROID_EMULATOR_IMAGE:-${defaultAndroidEmulatorImage}}`
-  const networkName = resolveComposeNetworkName(config)
-
-  const maybeAndroidEnv =
-    `      MCP_ANDROID_ENABLE: "1"\n      DOCKER_GIT_ANDROID_CONTAINER_NAME: "${androidContainerName}"\n      DOCKER_GIT_ANDROID_ADB_ENDPOINT: "\${DOCKER_GIT_ANDROID_ADB_ENDPOINT:-${androidContainerName}:5555}"\n      DOCKER_GIT_ANDROID_EMULATOR_IMAGE: "${androidImageRef}"\n`
-
-  const maybeAndroidService = `
-  ${config.serviceName}-android:
-    image: "${androidImageRef}"
-    container_name: ${androidContainerName}
-    privileged: true
-    environment:
-      EMULATOR_DEVICE: "\${DOCKER_GIT_ANDROID_DEVICE:-Samsung Galaxy S10}"
-      WEB_VNC: "\${DOCKER_GIT_ANDROID_WEB_VNC:-true}"
-      EMULATOR_HEADLESS: "\${DOCKER_GIT_ANDROID_HEADLESS:-true}"
-    devices:
-      - /dev/kvm
-    ports:
-      - "\${DOCKER_GIT_ANDROID_ADB_BIND_HOST:-127.0.0.1}:\${DOCKER_GIT_ANDROID_ADB_PORT:-5555}:5555"
-      - "\${DOCKER_GIT_ANDROID_NOVNC_BIND_HOST:-127.0.0.1}:\${DOCKER_GIT_ANDROID_NOVNC_PORT:-6080}:6080"
-${renderResourceLimits(resourceLimits)}    volumes:
-      - ${androidVolumeName}:/root/.android
-    networks:
-      - ${networkName}
-`
-
-  return {
-    maybeAndroidEnv,
-    maybeAndroidService,
-    maybeAndroidVolume: `  ${androidVolumeName}:`
-  }
-}
 
 const isResolvedComposeResourceLimits = (
   value: ResolvedComposeResourceLimits | ComposeResourceLimits
@@ -292,7 +176,10 @@ const buildComposeFragments = (
     resourceLimits.playwright,
     options
   )
-  const android = buildAndroidFragments(config, resourceLimits.playwright)
+  const android = buildAndroidFragments(
+    config,
+    renderResourceLimits(resourceLimits.playwright)
+  )
 
   return {
     networkMode,
